frontend_focused/backend/requirements.txt
asgiref==3.11.1
Deprecated==1.2.18
distlib==0.3.8
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework-camel-case==1.4.2
Faker==33.3.0
filelock==3.16.1
greenlet==3.1.0
lxml==5.3.1
msgpack==1.1.0
packaging==24.2
pikepdf==9.5.2
pillow==11.1.0
platformdirs==4.3.6
pynvim==0.5.0
sqlparse==0.5.5
typing_extensions==4.15.0
virtualenv==20.26.6
wrapt==1.17.2
